includes structure change

+++ icmp_funcs.cpp
@@ -1,13 +1,13 @@
 #include "icmp_funcs.h"
-#include <cassert>
-u_int16_t compute_icmp_checksum (const u_int16_t *buff, int length)
+
+short compute_icmp_checksum (const short *buff, int length)
 {
-	u_int32_t sum;
-	const u_int16_t* ptr = buff;
+	long sum;
+	const short* ptr = buff;
 	assert (length % 2 == 0);
 	for (sum = 0; length > 0; length -= 2)
 		sum += *ptr++;
 	sum = (sum >> 16) + (sum & 0xffff);
-	return (u_int16_t)(~(sum + (sum >> 16)));
+	return (short)(~(sum + (sum >> 16)));
 }
